% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acef_dplyr.R
\name{acef_dplyr}
\alias{acef_dplyr}
\title{ACEF Score Calculation with dplyr environment}
\arguments{
\item{data}{A data frame containing the columns: age, creatinine, and ejection_fraction.}

\item{...}{Additional arguments passed to dplyr functions.}
}
\value{
A data frame with an additional column for ACEF score.
}
\description{
This function calculates the ACEF score based on age, creatinine, and ejection fraction using dplyr.
}
