% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acef_ii_dplyr.R
\name{acef_ii_dplyr}
\alias{acef_ii_dplyr}
\title{ACEF II Score Calculation with dplyr environment}
\arguments{
\item{data}{A data frame containing the columns: age, creatinine, and ejection_fraction, emerjency_operation, hct.}

\item{...}{Additional arguments passed to dplyr functions.}
}
\value{
A data frame with an additional column for ACEF score.
}
\description{
This function calculates the ACEF II score based on age, creatinine, and ejection fraction, emerjency state of surger and hematocrit valus if inferor to 36 using dplyr.
}
