% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acef_ii.R
\name{acef_ii_score}
\alias{acef_ii_score}
\title{ACEF II Score Calculator}
\usage{
acef_ii_score(
  age = 60,
  creatinine = 2,
  ejection_fraction = 60,
  emergency_operation = FALSE,
  hct = 36
)
}
\arguments{
\item{age}{Age of the patient in years.}

\item{creatinine}{Serum creatinine level in mg/dL.}

\item{ejection_fraction}{Ejection fraction percentage}

\item{emergency_operation}{Logical value indicating if the operation is an emergency.}

\item{hct}{Hematocrit level in percentage below 36 %, positive number.}
}
\value{
ACEF score in numbers.
}
\description{
This function calculates the ACEF II score based on age, creatinine, emergency operation, HCT (hematocrit bellwo 36 %) and ejection fraction.
}
\details{
ACEF II = Age/EF + 2.0 (if serum creatinine > 2.0 mg/dL) + 3.0 (if emergency surgery) + 0.2 HCT points below 36%
}
\examples{
acef_score(age = 65, creatinine = 1.2, ejection_fraction = 55)
}
