% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acef.R
\name{acef_score}
\alias{acef_score}
\title{ACEF Score Calculator}
\usage{
acef_score(age, creatinine, ejection_fraction)
}
\arguments{
\item{age}{Age of the patient in years.}

\item{creatinine}{Serum creatinine level in mg/dL.}

\item{ejection_fraction}{Ejection fraction percentage.}
}
\value{
ACEF score in numbers.
}
\description{
This function calculates the ACEF score based on age, creatinine, and ejection fraction.
}
\examples{
acef_score(age = 65, creatinine = 1.2, ejection_fraction = 55)
}
