% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euroscore_ii_dplyr.R
\name{es_ii}
\alias{es_ii}
\title{EuroSCORE II}
\usage{
es_ii(data, ...)
}
\arguments{
\item{data}{A data frame containing the required variables.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame with the calculated EuroSCORE II risk score.
}
\description{
This function calculates the EuroSCORE II risk score for patients undergoing cardiac surgery.
It uses the dplyr package for data manipulation.
}
